Split ProductUpgraderInfo into a shared and non-shared class
This will facilitate adding tests for the logic that the hooks projects do not
depend on, as those projects do not support the PhysicalFileSystem abstraction.
