+++ GVFS/GVFS.Hooks/GVFS.Hooks.Windows.csproj
@@ -94,8 +94,8 @@
     <Compile Include="..\GVFS.Common\ProcessResult.cs">
       <Link>Common\ProcessResult.cs</Link>
     </Compile>
-    <Compile Include="..\GVFS.Common\ProductUpgraderInfo.cs">
-      <Link>Common\ProductUpgraderInfo.cs</Link>
+    <Compile Include="..\GVFS.Common\ProductUpgraderInfo.Shared.cs">
+      <Link>Common\ProductUpgraderInfo.Shared.cs</Link>
     </Compile>
     <Compile Include="..\GVFS.Common\Tracing\EventLevel.cs">
       <Link>Common\Tracing\EventLevel.cs</Link>
